Adding windows-latest to the matrix strategy.
That required to link against the compiled version of boost unit-test
framework.
To use --toolchain to tell cmake about the toolchain because
working directory seemed to be different in linux vs. windows.
Lastly, I changed the step that executes unit tests so that it doesn't use the
test target because it does not exist in windows (it is called RUN_TESTS).